Rename files to accurately reflect class name and purpose
