+++ test/rotation.rb
